Организовывает модульную структуру JS
